Add Gif Thumbnail, Load Posts & Fix Loading Times
Added a Thumbnail for gifs like for videos, with a distinct functionality that clicking on a rendered gif can toggle back to the thumbnail and vice versa.
The Default limit of posts to be rendered is set at 20, with each click on the button loading 20 more posts. All of these changes will significantly lower loading times and provide a better experience for the user.